add functions to change box colors on mouseover
Created a mouseover event listener for each box that runs a function
to see if the box/div has a certain class, if it does it will add different class
that is defined in the css. this is able to change the box color multiple times
by adding additional if statements and css classes.