add functionality to grid size slider and color dropdown
Added functions to change the grid size based on the slider, this also calls another function
to clear the grid divs and event listeners and initializes a new play grid.
Added functions to change the color changing aspect through the drop down. this funtion
does not refresh the play grid, as to give functionality to the eraser.